Optimize the file distribution for better dev
we create Utility and Operation to sever load and save model.
The Utility is to sever Operation at the same time.
And several test, we are sure that the function para_m and para_m_pos are passing by reference, what they have change is saved in load_save_model.py